Update highlightRegionTest.cpp
Update highlightRegionTest.cpp

+++ test/highlightRegionTest.cpp
@@ -9,7 +9,7 @@ void subimage2Test(const std::size_t sizex = 3, const std::size_t sizey = 3)
 {
     TinyDIP::Image<T> image1(sizex, sizey);
     image1.setAllValue(10);
-    image1.at(2, 2) = 1;
+    image1.at(static_cast<std::size_t>(2), static_cast<std::size_t>(2)) = 1;
     std::cout << "Width: " + std::to_string(image1.getWidth()) + "\n";
     std::cout << "Height: " + std::to_string(image1.getHeight()) + "\n";
     auto image2 = TinyDIP::copyResizeBicubic<T>(image1, 12, 12);
